Simulation Code/spaceRobot_RL/requirements.txt
git
glfw
gym
ipykernel
ipython
jupyter
Keras-Preprocessing==1.1.2
Markdown
matplotlib==3.3.0
mujoco-py==2.0.2.11
numba
numpy
pandas
plotly
pyglet
PyOpenGL
PyYAML
scikit-learn
scipy
seaborn
tensorboard
tensorflow-gpu
torch
tqdm
